// apps/web/src/components/ShareImageCard.d.ts
import React from 'react';
import type { GameState } from '@republica/game-engine';
interface Props {
    gameState: GameState;
    presidentId: string;
    score: number;
}
export declare const ShareImageCard: React.ForwardRefExoticComponent<Props & React.RefAttributes<HTMLDivElement>>;
export {};
//# sourceMappingURL=ShareImageCard.d.ts.map